:truck: split audioware types
