+++ audioware/src/types/audioware.rs
@@ -1,25 +1,14 @@
 use std::time::Duration;
 
 use kira::tween::{Easing, Tween};
-use red4ext_rs::{class_kind::Scripted, log, types::Ref, NativeRepr, PluginOps, ScriptClass};
+use red4ext_rs::{class_kind::Scripted, log, types::Ref, PluginOps, ScriptClass};
 
 use crate::Audioware;
 
 mod subtitles;
 pub use subtitles::propagate_subtitles;
-
-#[allow(clippy::enum_variant_names, dead_code)]
-#[derive(Debug, Default, Clone, Copy, PartialEq, Eq, PartialOrd, Ord)]
-#[repr(i64)]
-pub enum AudiowareEasing {
-    #[default]
-    InPowf = 0,
-    OutPowf = 1,
-    InOutPowf = 2,
-}
-unsafe impl NativeRepr for AudiowareEasing {
-    const NAME: &'static str = "AudiowareEasing";
-}
+mod easing;
+pub use easing::*;
 
 #[derive(Debug)]
 #[repr(C)]
